+++ goout-server/src/register/register.service.ts
@@ -6,34 +6,5 @@ import { RedisService } from "src/util/redis.service";
 
 @Injectable()
 export class RegisterService {
-  constructor(
-    private readonly userdataservice: UserdataService,
-    private readonly redisservice: RedisService
-  ) {}
-
-  async send_AuthNum(req: RegisterDataDto) {
-    try {
-      let authNum = Math.random().toString().substr(2, 6);
-
-      const smtpTransport = nodemailer.createTransport({
-        service: "Gmail",
-        auth: {
-          user: process.env.NODEMAILER_USER,
-          pass: process.env.NODEMAILER_PASS,
-        },
-      });
-
-      const mailOptions = {
-        from: process.env.NODEMAILER_USER,
-        to: req.email,
-        subject: "Go-Out 회원가입 E-Mail인증번호",
-        text: `인증번호는 ${authNum}입니다.`,
-      };
-      let created_data = await this.userdataservice.createUserdata(req);
-      this.redisservice.add_redis(created_data.userid,authNum);
-    } catch (error) {
-      console.log(error);
-      throw new HttpException("이메일 전송 에러", HttpStatus.CONFLICT);
-    }
-  }
+  
 }
